requirements.txt
gradio==5.12.0
gtts==2.5.4
pydub==0.25.1
speechrecognition==3.13.0
groq==0.15.0
python-dotenv==1.0.1
pillow==11.1.0
sentence-transformers==2.2.2